requirements.txt
discord
python-dotenv
pyyaml